refactor: move compile_builder and _execute_cascade to run domain
Move _get_artifacts_list and compile_builder from domain/blueprint/service
to a new domain/run/compile.py. domain/blueprint/service now only owns
validation/expansion and blueprint persistence; it no longer depends on
earthkit.workflows or fiab_core.artifacts.

Move ProductToOutputId and _execute_cascade (renamed execute_cascade) from
domain/run/service to domain/run/cascade.py, alongside encode_result.
domain/run/cascade.py now owns all Cascade submission and result encoding.

domain/run/service.py is simplified: it imports compile_builder from
domain/run/compile and execute_cascade / ProductToOutputId from
domain/run/cascade.

routes/run.py updated to import ProductToOutputId from domain/run/cascade.
